clean up code and fix colorscheme
