+++ frontend/src/common/node.ts
@@ -109,22 +109,9 @@ export const mapCorrelationToColor = (correlation: number) => {
 
 export const isSubgoal = (node: GraphNode): boolean => node.type === 'Delmål';
 
-export const isU4sscKPI = (node: GraphNode): boolean => node.type === 'U4SSC KPI';
-export const isOecdKPI = (node: GraphNode): boolean => node.type === 'OECD KPI';
-
-export const isKpiSelected = (
-  node: GraphNode,
-  filter: KpiFilter,
-  // radioSelected: number,
-): boolean => {
-  if (node.type === 'U4SSC KPI' && !filter.u4ssc) {
-    console.log(node.type, filter.oecd);
-    return true;
-  }
-  if (node.type === 'OECD KPI' && !filter.oecd) {
-    console.log(node.type, filter.oecd);
-    return true;
-  }
+export const isKpiSelected = (node: GraphNode, filter: KpiFilter): boolean => {
+  if (node.type === 'U4SSC KPI' && !filter.u4ssc) return true;
+  if (node.type === 'OECD KPI' && !filter.oecd) return true;
   return false;
 };
